Final touches to translation editor, fixes for over-escaping and preventing XSS, provide english samples, translation bugfixes

+++ TW.Vault/Configuration.cs
@@ -153,6 +153,6 @@ namespace TW.Vault
 
     public class TranslationConfiguration
     {
-        public int BaseTranslationId { get; set; }
+        public short BaseTranslationId { get; set; }
     }
 }
